PROJECT 1X4G Page
Added page that with this text:

We can replace the DNA in any virus to add DNA to a cell for example If we replaced the DNA inside or added new genes to a bacteriophage could we create super bacteria with extremely advanced DNA? Another thing I was wondering if that could you use a bacteriophage to modify a mitochondria because mitochondria can self-replicate and are about the size of a cell couldn't you use a bacteriophage to add genes to the DNA of the part of the cell thus genetically modifying a human on a smaller scale now. You could use a Bacteria and bacteriophage to add new DNA to old peoples DNA to allow them to regenerate because as you age DNA splits also, I believe Dolly the sheep died because it only had DNA from 1 parent meaning the DNA strands were half the size so it would have increased aging and missing details needed for life. If we change carbohydrate chains then the virus will not be able to attach to the cell to create immunity to all types of viruses until they evolve. We can grow meat and infect with the virus that we have put our new DNA in to mass-produce it. If we combine my virus that inserts DNA with Crispr-Cas9 we can add and replace DNA.

We could use a system that has a neuromorphic chip attached and a CPU (Von Neumann architecture) as well as a quantum chip connected up to neurons grown to interact with the different types of chips this could create a thing that could think and process in every way possible a ‘SUPER BRAIN’. I want to grow a rack of neuron-controlled chips and connect them all up to create a hive of mini-brains. Also, the brain can communicate with an output which is amplified current from the neurons like TDCS a device but sending brain electrical signals and also use brain electrical scans to send signals to the neurons to allow it to learn how to interact with the brain because it creates a input and output to the brain.